+++ tests/test_project.py
@@ -1,7 +1,7 @@
 # ---------------------------------------------------------------------
 # Gufo Labs: Project structure tests
 # ---------------------------------------------------------------------
-# Copyright (C) 2022-24, Gufo Labs
+# Copyright (C) 2022-25, Gufo Labs
 # See LICENSE.md for details
 # ---------------------------------------------------------------------
 
@@ -50,11 +50,17 @@ def _get_project_info() -> Tuple[str, str]:
         assert len(d) == 1
         return d[0]
 
+    def has_init(*args: str) -> bool:
+        return Path(*(*args, "__init__.py")).exists()
+
     ns = explore_dir(ROOT, "src")
     if ns == "gufo":
         # gufo.* namespace
         pkg = explore_dir(ROOT, "src", ns)
-        return os.path.join("src", ns, pkg), f"{ns}.{pkg}"
+        if has_init("src", ns, pkg):
+            return os.path.join("src", ns, pkg), f"{ns}.{pkg}"
+        pkg2 = explore_dir(ROOT, "src", ns, pkg)
+        return os.path.join("src", ns, pkg, pkg2), f"{ns}.{pkg}.{pkg2}"
     return os.path.join("src", ns), ns
 
 
@@ -68,9 +74,6 @@ REQUIRED_FILES = [
     ".github/ISSUE_TEMPLATE/feature-request.yml",
     ".github/PULL_REQUEST_TEMPLATE.md",
     ".gitignore",
-    ".requirements/docs.txt",
-    ".requirements/lint.txt",
-    ".requirements/test.txt",
     "CHANGELOG.md",
     "CITATION.cff",
     "CODE_OF_CONDUCT.md",